Добавляет рефакторинк кода при помощи импорта и экспорта
